Added Filter cache
The filter cache is a simple layer which caches the return of a filter, also watches and caches all the changes which are made on the Request object.

It does not implement any cache mechanism by default, but it is fairly simple by implement an interface which is generated along with Route class.

Format:

```php
    /** @Filter <name> @Cache <ttl> */
```

The <ttl> value is passed to the Cache object, so we assumed that this value won't be ignored.